+++ src/Utilities/AresFunctions.h
@@ -45,10 +45,7 @@ public:
 	static void* (__thiscall* GetTunnel)(void*, HouseClass*);
 	static void(__thiscall* AddPassengerFromTunnel)(void*, BuildingClass*, FootClass*);
 private:
-
-	static constexpr bool _maybe = false;
-
-	static constexpr bool AresWasWrongAboutSpawnSurvivors = _maybe;
+	static constexpr bool AresWasWrongAboutSpawnSurvivors = false;
 
 	static void* _SWTypeExtMap;
 	static AresSWTypeExtData* (__thiscall* _SWTypeExtMapFind)(void*, SuperWeaponTypeClass*);